# Documentation/OpenAPI-Hotel-ContentAPI-3.0.yaml
openapi: 3.0.1
info:
  title: Hotel Content API
  description: |
    # APItude Hotel Content API

    Hotel Content API has been designed to retrieve **all the details** of the hotels available in our [Hotel Booking API](/documentation/hotels/booking-api/api-reference/). 

    Hotel Content API contains several methods that allow you to obtain the details of the hotels as well as list of destinations, countries and the master types that complement the hotels informatoin (like room type descriptions, accommodation type descriptions, categories descriptions...).

    Note that the Languages operation is connected to all the operations in which a description is returned, since we return the language code in which that description has been returned.
  termsOfService: https://developer.hotelbeds.com/api-terms-use
  contact:
    name: Apitude Support
    url: https://developer.hotelbeds.com
    email: apitude@hotelbeds.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
...
servers:
- url: https://api.test.hotelbeds.com/hotel-content-api/{version}
  variables:
    version:
      default: "1.0"
      description: Specification version
...
paths:
...
  /locations/countries:
    get:
      tags:
      - Locations
      summary: Countries
      description: |-
        Returns the list of available countries.

        This is the higher level available in our destination tree. We do not allow search by country, but it's included for organizational reasons (to group destinations) and because the country codes returned are used in other levels of APItude (for example to define the sourceMarket in Availability operation of [Hotel Booking API](/documentation/hotels/booking-api/api-reference/).
      operationId: countriesUsingGET
      parameters:
      - $ref: '#/components/parameters/CommonRequestHeaderAccept'
      - $ref: '#/components/parameters/CommonRequestHeaderAcceptEncoding'
      - $ref: '#/components/parameters/CommonRequestParameterFields'
      - $ref: '#/components/parameters/CommonRequestParameterCodes'
      - $ref: '#/components/parameters/CommonRequestParameterLanguage'
      - $ref: '#/components/parameters/CommonRequestParameterFrom'
      - $ref: '#/components/parameters/CommonRequestParameterTo'
      - $ref: '#/components/parameters/CommonRequestParameterUseSecondaryLanguage'
      - $ref: '#/components/parameters/CommonRequestParameterLastUpdateTime'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCountriesRS'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiCountriesRS'
        400:
          description: Bad Request
          content: {}
        401:
          description: Unauthorized
          content: {}
        402:
          description: Payment Required
          content: {}
        500:
          description: Internal Server Error
          content: {}
        503:
          description: Product Exception/Service Unavailable
          content: {}
  /locations/destinations:
    get:
      tags:
      - Locations
      summary: Destinations
      description: |-
        Return the list of available destinations, zones and grouping zones. 

        Destinations are the highest level entities in our destination tree by which availability can be searched. Please note that destination codes are not IATA codes. It is very important to understand that our 'destination' will not always only correspond with a unique city, but usually with the main city and the area surrounding the city. 

        For example, the destination code NYC in the US, corresponds with the whole New York Area, so it is not limited to the city of New York, but also covers the surrounding areas, such as New Jersey and The Hamptons. Another example is BCN in Spain, which corresponds with the entire Barcelona Province; so it searches hotels located in the city of Barcelona, but also in other cities within the Province of Barcelona, such us Hospitalet de Llobregat, Terrassa and Sabadell. So when you search for availability using the destination code, you will receive results that may be located far from the main city for which you are actually searching.

        It is very important to check the locations that the search retrieves to avoid booking hotels in a different location than the one requested by your customers.
      operationId: destinationsUsingGET
      parameters:
      - $ref: '#/components/parameters/CommonRequestHeaderAccept'
      - $ref: '#/components/parameters/CommonRequestHeaderAcceptEncoding'
      - $ref: '#/components/parameters/CommonRequestParameterFields'
      - name: codes
        in: query
        description: 'Filter to limit the results for an specific destination or list
          of destinations.'
        required: false
        style: form
        explode: false
        schema:
          type: array
          items:
            type: string
      - name: countryCodes
        in: query
        description: 'Filter for a specific country of list of countries.'
        required: false
        style: form
        explode: false
        schema:
          type: array
          items:
            type: string
      - $ref: '#/components/parameters/CommonRequestParameterLanguage'
      - $ref: '#/components/parameters/CommonRequestParameterFrom'
      - $ref: '#/components/parameters/CommonRequestParameterTo'
      - $ref: '#/components/parameters/CommonRequestParameterUseSecondaryLanguage'
      - $ref: '#/components/parameters/CommonRequestParameterLastUpdateTime'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDestinationsRS'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiDestinationsRS'
        400:
          description: Bad Request
          content: {}
        401:
          description: Unauthorized
          content: {}
        402:
          description: Payment Required
          content: {}
        500:
          description: Internal Server Error
          content: {}
        503:
          description: Product Exception/Service Unavailable
          content: {}
...
components:
  parameters:
    CommonRequestHeaderAccept:
      name: Accept
      in: header
      description: 'Response format, posible values are application/json and application/xml.'
      required: true
      schema:
        type: string
        enum: [application/xml, application/json]
    CommonRequestHeaderAcceptEncoding:
      name: Accept-Encoding
      in: header
      description: 'Defines type of encoding used on the data.'
      required: true
      schema:
        type: string
        default: gzip
    CommonRequestParameterFields:
      name: fields
      in: query
      description: 'The list of fields to be received in the response. To retrieve all the fields use ‘all’. If nothing is specified, all fields are returned. See the complete list of available fields in the response.'
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
          example: all
    CommonRequestParameterCodes:
      name: codes
      in: query
      description: 'List of codes to filter by.'
      required: false
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
    CommonRequestParameterLanguage:
      name: language
      in: query
      description: 'The language code for the language in which you want the descriptions to be returned. If language is not specified, English will be used as default language.'
      required: false
      schema:
        type: string
      example: CAS
    CommonRequestParameterFrom:
      name: from
      in: query
      description: 'The number of the initial record to receive. If nothing is specified, 1 is the default value.'
      required: false
      schema:
        type: integer
        format: int32
        example: 1
    CommonRequestParameterTo:
      name: to
      in: query
      description: 'The number of the final record to receive. If nothing is indicated, 100 is the default value.'
      required: false
      schema:
        type: integer
        format: int32
        example: 100
    CommonRequestParameterUseSecondaryLanguage:
      name: useSecondaryLanguage
      in: query
      description: 'Defines if you want to receive the descriptions in English if the description is not available in the language requested.'
      required: false
      schema:
        type: boolean
    CommonRequestParameterLastUpdateTime:
      name: lastUpdateTime
      in: query
      description: 'Specifying this parameter limits the results to those modified or added after the date specified. The required format is YYYY-MM-DD.'
      required: false
      schema:
        type: string
      example: '2020-03-03'
...
  schemas:
...
    ApiCountriesRS:
      type: object
      description: 'Response of Countries operation.'
      properties:
        auditData:
          $ref: '#/components/schemas/ApiAuditData'
        countries:
          type: array
          description: 'List of countries.'
          items:
            $ref: '#/components/schemas/ApiCountry'
        from:
          type: integer
          description: 'Lowest value of the range of returned countries.'
          format: int32
        to:
          type: integer
          description: 'Highest value of the range of returned countries.'
          format: int32
        total:
          type: integer
          description: 'Total number of countries matching the parameters used in the request (regardless of pagination/returned range).'
          format: int32
...
    ApiDestination:
      type: object
      description: 'Information on a destination of the destination tree.'
      properties:
        code:
          type: string
          description: 'Code of the destination.'
          example: PMI
        countryCode:
          type: string
          description: 'Hotelbeds code for the country'
          example: ES
        groupZones:
          type: array
          description: 'List of groups of zones.'
          items:
            $ref: '#/components/schemas/ApiGroupZone'
        isoCode:
          type: string
          description: 'ISO Country Code.'
          example: ES
        name:
          $ref: '#/components/schemas/ApiContent'
        zones:
          type: array
          description: 'List of zones in which the destination is divided.'
          items:
            $ref: '#/components/schemas/ApiZone'
    # ApiCodedElement:
      # type: object
      # properties:
        # code:
          # type: string
        # description:
          # $ref: '#/components/schemas/ApiContent'
...
    ApiGroupZone:
      type: object
      description: 'Information on a group of zones'
      properties:
        groupZoneCode:
          type: string
          description: 'Code of the group of zones.'
          example: MALLOPMI16
        name:
          $ref: '#/components/schemas/ApiContent'
        zones:
          type: array
          description: 'List of zones that belong to the group.'
          items:
            type: integer
            format: int32
...
    ApiZone:
      type: object
      description: 'Information on the zone inside the destination where to search for hotels.'
      properties:
        description:
          $ref: '#/components/schemas/ApiContent'
        name:
          type: string
          description: 'Internal zone name. (only available for Hotel Detail)'
          example: Candabong
        zoneCode:
          type: integer
          description: 'Internal Zone Code.'
          format: int32
...
    ApiDestinationsRS:
      type: object
      description: 'Response of Destinations operation.'
      properties:
        auditData:
          $ref: '#/components/schemas/ApiAuditData'
        destinations:
          type: array
          description: 'List of available destinations.'
          items:
            $ref: '#/components/schemas/ApiDestination'
        from:
          type: integer
          description: 'Lowest value of the range of returned destinations.'
          format: int32
        to:
          type: integer
          description: 'Highest value of the range of returned destinations.'
          format: int32
        total:
          type: integer
          description: 'Total number of destinations matching the parameters used in the request (regardless of pagination/returned range).'
          format: int32
...
    ApiState:
      type: object
      description: 'State information.'
      properties:
        code:
          type: string
          description: 'Internal code of the state.'
          example: CA
        name:
          type: string
          description: 'Name of the state.'
          example: CALIFORNIA
...
    ApiAuditData:
      type: object
      description: 'Relevant internal information.'
      properties:
        environment:
          type: string
          description: 'Environment where the request has been sent and processed.'
          example: [awseuwest1, awseuwest1b, ip_10_185_91_142]
        # expiration:
          # type: string
          # deprecated: true
        # internal:
          # type: string
          # deprecated: true
        processTime:
          type: string
          description: 'Server process time in milliseconds.'
          example: '53'
        release:
          type: string
          description: 'Release version identifier.'
          example: 19117e1a462f1a1d7fa2fcf6371e3bbdb0c14a89
        requestHost:
          type: string
          description: 'IP address of the system host.'
          example: 10.185.84.47
        serverId:
          type: string
          description: 'Server code or ID (for internal use).'
          example: ip-10-185-91-142.eu-west-1.compute.internal.node.int-hbg-aws-eu-west-1.discovery
        timestamp:
          type: string
          description: 'GMT Date/time when the request has been processed.'
          example: 2020-03-03 13:22:20.821
...
    ApiCountry:
      type: object
      description: 'Country information.'
      properties:
        code:
          type: string
          description: 'Hotelbeds internal country code.'
          example: PT
        description:
          $ref: '#/components/schemas/ApiContent'
        isoCode:
          type: string
          description: 'ISO 3166-2 Country Code.'
          example: PT
        states:
          type: array
          description: 'List of states/provinces/territories belonging to the country.'
          items:
            $ref: '#/components/schemas/ApiState'
    ApiContent:
      type: object
      properties:
        content:
          type: string
          description: 'Text description, provided in the language defined in LanguageCode field.'
        languageCode:
          type: string
          description: 'Code of the language in which the description is returned.'
...
  securitySchemes:
    Apikey:
      type: apiKey
      name: Api-key
      description: Your registered API key
      in: header
    Xsignature:
      type: apiKey
      name: X-Signature
      description: SHA256 encoding signature
      in: header
security:
  - Apikey: []
    Xsignature: []